Drastically improve waterfall performance
set_at is extremely slow, plus we're doing a call to clamp() and
color_func() once per pixel. This causes the waterfall to be a bit
underwhelming. Using surfarray makes it quicker - I think you can
probably do away with the pixel iterator completely and copy the entire
row in one go (equivalent of a C memcpy())